+++ native/src/base/lib.rs
@@ -11,12 +11,14 @@ use cxx_extern::*;
 pub use files::*;
 pub use logging::*;
 pub use misc::*;
+pub use result::*;
 
 mod cstr;
 mod cxx_extern;
 mod files;
 mod logging;
 mod misc;
+mod result;
 mod xwrap;
 
 #[cxx::bridge]
